refactor: 헬스 체크 api를 코어에서 이동
